requirements.txt
# Web & Backend
flask==3.0.2
flask-cors==4.0.0

# Data Streaming (Kafka)
kafka-python==2.0.2

# Database
psycopg2-binary

# AI & Machine Learning
tensorflow>=2.16.1
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4

# Email & Utilities
python-dotenv==1.0.1